check ollama status every 3s
