+++ gui/src/components/OnboardingCard/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from "./useCheckOllamaModels";
 export * from "./useSubmitOnboarding";
+export * from "./useOnboardingCard";
